Refs #3753. Moved serial protocol file to transport profile.
