+++ include/uxr/client/profile/transport/serial/serial_protocol.h
@@ -12,17 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _UXR_CLIENT_CORE_COMMUNICATION_SERIAL_PROTOCOL_H_
-#define _UXR_CLIENT_CORE_COMMUNICATION_SERIAL_PROTOCOL_H_
-
+#ifndef _UXR_CLIENT_PROFILE_TRANSPORT_SERIAL_SERIAL_PROTOCOL_INTERNAL_H_
+#define _UXR_CLIENT_PROFILE_TRANSPORT_SERIAL_SERIAL_PROTOCOL_INTERNAL_H_
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
 #include <stdint.h>
-#include <stddef.h>
-#include <stdbool.h>
 
 #define UXR_FRAMING_BEGIN_FLAG 0x7E
 #define UXR_FRAMING_ESC_FLAG 0x7D
@@ -58,34 +55,8 @@ typedef struct uxrSerialIO
 
 } uxrSerialIO;
 
-void uxr_init_serial_io(uxrSerialIO* serial_io, uint8_t local_addr);
-
-void update_crc(uint16_t* crc, const uint8_t data);
-bool get_next_octet(uxrSerialIO* serial_io, uint8_t* octet);
-bool add_next_octet(uxrSerialIO* serial_io, uint8_t octet);
-
-struct uxrSerialPlatform;
-typedef size_t (*uxr_write_cb)(struct uxrSerialPlatform*, uint8_t*, size_t, uint8_t*);
-typedef size_t (*uxr_read_cb)(struct uxrSerialPlatform*, uint8_t*, size_t, int, uint8_t*);
-
-size_t uxr_write_serial_msg(uxrSerialIO* serial_io,
-                            uxr_write_cb write_cb,
-                            void* cb_arg,
-                            const uint8_t* buf,
-                            size_t len,
-                            uint8_t remote_addr,
-                            uint8_t* errcode);
-size_t uxr_read_serial_msg(uxrSerialIO* serial_io,
-                           uxr_read_cb read_cb,
-                           void* cb_arg,
-                           uint8_t* buf,
-                           size_t len,
-                           uint8_t* remote_addr,
-                           int timeout,
-                           uint8_t* errcode);
-
 #ifdef __cplusplus
 }
 #endif
 
-#endif //_UXR_CLIENT_CORE_COMMUNICATION_SERIAL_PROTOCOL_H_
+#endif //_UXR_CLIENT_PROFILE_TRANSPORT_SERIAL_SERIAL_PROTOCOL_INTERNAL_H_
